refactor: 테스트 클래스 패키지 정리
- 프로덕션 클래스 패키지 구조와 동일하도록 테스트 클래스의 패키지 경로 수정
